implement clue generation
This commit implements the generation of clues using their
various forms. The results are _correct_ in the sense that
the generated clues are all valid, but we just keep
generating clues randomly until the slots are all filled.
This has a few problems:

  - there are too many clues for some tiles; this makes them
    easy to figure out, and the clues are redundant (we may
    even have duplicate clues!)

  - some tiles may not get a clue; because of the
    redundancy, even with 24 clue slots, we may fail to
    mention sets of tiles at all, leading to unsolvable
    puzzles

As a side note, the clue generation depends on all rows
having the same number of columns. In theory the puzzle
could be ragged, and the rest of the code supports that.
If a puzzle can violate that assumption, it makes clue
generation rather tricky (e.g., if you pick column X in a
row such that it is not on the far right side, you know that
X+1 is valid in _that_ row, but not necessarily in other
rows).