allow toggling active/inactive clues
I find this feature key for keeping track of which clues are
still relevant. The diff's a bit big here, because we have
to refactor the clues to actually store their data (so they
can display and clear later). We probably needed to do this
eventually, anyway, as the clues need to know what they
"mean" in order to test-solve the puzzle with them.

The implementation violates some of the existing OO design:

  - the "toggle" flag for clues is a global, as it's getting
    to be a real pain to point to "the" puzzle. It's
    probably worth declaring that there is a singleton
    puzzle, and removing some of the object-passing
    boilerplate.

  - the clue objects should probably all inherit so they can
    share the toggle/display code. But here I just use
    external functions, which is a bit hacky.