fix race on row.removePossible()
When we know that slot X has some value, we remove it as a possibility
for all the other slots in that row. But we do so by calling
slot.discard(), which does more than we want. After removing the
possibility, it calls row.checkSingleton() which will check _all_ slots
in the row. This is wrong because it will look at slots for which we
haven't yet removed the possibility. Worse, because we _did_ remove the
possibility from the slot that actually checked it, we'll mistakenly
think that some other slot is the only option!

We can fix this by pulling the "remove possibility" parts out of the
discard function, and calling them separately from row.removePossible().