don't checkSingleton() for already-solved slots
Once a slot has a value, there's no point in seeing if it has only one
possible option left. This is not only wasteful, but it also may cause
us to complain about the move.

Arguably we should be removing all possibilities when it becomes solved,
but this works equally well.